cleaning up namespacing

+++ App/controllers/pwr_ctrl.h
@@ -1,6 +1,8 @@
 #ifndef __POWER_CONTROLLER_H__
 #define __POWER_CONTROLLER_H__
 
+#include "stm32f4xx_hal.h"
+
 typedef struct
 {
     // cfg
@@ -9,8 +11,8 @@ typedef struct
     uint16_t mosfet_pin;
     // state
     uint16_t mcp_target_voltage;
-    double cur_voltage;
-    double cur_current;
+    float cur_voltage;
+    float cur_current;
     bool output_enabled;
 } Power_Controller_t;
 
@@ -26,7 +28,7 @@ Power_Status_t Power_Init(Power_Controller_t *ctrl, I2C_HandleTypeDef *i2c,
 Power_Status_t Power_SetVoltage(Power_Controller_t *ctrl, uint16_t value);
 Power_Status_t Power_Enable(Power_Controller_t *ctrl, bool enable);
 Power_Status_t Power_Update(Power_Controller_t *ctrl);
-Power_Status_t Power_Read(Power_Controller_t *ctrl, double *voltage, double *current);
+Power_Status_t Power_Read(Power_Controller_t *ctrl, float *voltage, float *current);
 void Power_SweepRange(Power_Controller_t *ctrl);
 
 #endif
